// dashboard/src/utils/hasura-api/generated/schemas/getEventAndInvocationLogsByIdArgs.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */

export interface GetEventAndInvocationLogsByIdArgs {
  /** ID of the event */
  event_id: string;
  /** Name of the source database of the trigger */
  source?: string;
  /** Maximum number of invocation logs to be returned in one API call */
  invocation_log_limit?: number;
  /** Starting point from where the invocation logs need to be returned */
  invocation_log_offset?: number;
}