// dashboard/src/utils/hasura-api/generated/schemas/getEventAndInvocationLogsByIdOperation.ts
/**
 * Generated by orval v8.0.2 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { GetEventAndInvocationLogsByIdArgs } from './getEventAndInvocationLogsByIdArgs';
import type { GetEventAndInvocationLogsByIdOperationType } from './getEventAndInvocationLogsByIdOperationType';

export interface GetEventAndInvocationLogsByIdOperation {
  type: GetEventAndInvocationLogsByIdOperationType;
  args: GetEventAndInvocationLogsByIdArgs;
}